Trial Nav With Maps

+++ src/threonine/midlayer/MapGraphicGetParam.java
@@ -7,6 +7,6 @@ public class MapGraphicGetParam {
     public long universeid = 0;
     public long subsetid = 0;
     public long trialid = 0;
-    public int nodecode = 0;
+    public long nodecode = 0;
 }
 //**************************************************************************
